extract categories and templates autocomplete
